requirements (2).txt
gradio
google-generativeai
chardet
